+++ src/blog/Blog.tsx
@@ -8,9 +8,8 @@ import BlogHero from "./components/hero/BlogHero";
 import PostPreview from "src/blog/components/postPreview/PostPreview";
 import PostPreviewGrid from "./components/postPreviewGrid/PostPreviewGrid";
 import LoadingNews from "./components/loadingNews/LoadingNews";
-import CustomErrorMessage, {
-  homeLink,
-} from "./components/customErrorMessage/CustomErrorMessage";
+import CustomErrorMessage from "./components/customErrorMessage/CustomErrorMessage";
+import { homeLink } from "./components/utils/linkTypes";
 
 interface BlogProps {
   blog: BlogPage;
