+++ src/blog/components/customErrorMessage/CustomErrorMessage.tsx
@@ -1,26 +1,8 @@
 import Button from "src/components/buttons/Button";
 import LinkButton from "src/components/linkButton/LinkButton";
 import Text from "src/components/text/Text";
-import { ILink, LinkType } from "studio/lib/payloads/navigation";
+import { ILink } from "studio/lib/payloads/navigation";
 import styles from "./customErrorMessage.module.css";
-
-export const homeLink = {
-  _key: "return-home",
-  _type: "link",
-  linkTitle: "Return to home",
-  linkType: LinkType.Internal,
-  internalLink: {
-    _ref: "/",
-  },
-};
-
-export const studioLink = {
-  _key: "go-to-studio",
-  _type: "link",
-  linkTitle: "Go to studio",
-  linkType: LinkType.Internal,
-  internalLink: { _ref: "/studio" },
-};
 
 interface CustomErrorMessageProps {
   title: string;
