+++ src/blog/components/informationMessage/InformationMessage.tsx
@@ -1,27 +1,9 @@
 import Button from "src/components/buttons/Button";
 import LinkButton from "src/components/linkButton/LinkButton";
 import Text from "src/components/text/Text";
-import { ILink, LinkType } from "studio/lib/payloads/navigation";
+import { ILink } from "studio/lib/payloads/navigation";
 import styles from "./informationMessage.module.css";
 import React from "react";
-
-export const homeLink = {
-  _key: "return-home",
-  _type: "link",
-  linkTitle: "Return to home",
-  linkType: LinkType.Internal,
-  internalLink: {
-    _ref: "/",
-  },
-};
-
-export const studioLink = {
-  _key: "go-to-studio",
-  _type: "link",
-  linkTitle: "Go to studio",
-  linkType: LinkType.Internal,
-  internalLink: { _ref: "/studio" },
-};
 
 interface InformationMessageProps {
   title: string;
